code cleanup

+++ selektor/data.py
@@ -182,7 +182,7 @@ class WordVecs(object):
         """
         vocab_size = len(word_vecs)
         word_idx_map = dict()
-        W = np.zeros(shape=(vocab_size+1, k))            
+        W = np.zeros(shape=(vocab_size+1, k), dtype="float32")            
         W[0] = np.zeros(k)
         i = 1
         for word in word_vecs:
